refactor: extract search hook and split ui
